feat: add `toolbox-recreate` script to recreate a toolbox
Add `toolbox-recreate` which will tear down and create a new
toolbox, either one named on the command line or the default one.
This solves a problem with how to update a toolbox -- instead of
updating the container, just destroy it and recreate it with a
fresh image.

The list of packages to install in each toolbox is embedded in the
script itself and should be updated regularly.