feat: add `precmd` module for commands prior to evaluating prompt
Add a `precmd` module that contains commands that are executed
each time the prompt is evaluated, and is itself evaluated before
any other sections.  If the shell is not able to do command
substitution in the shell prompt, then this module is a no-op.

The `precmd` module sets a variable `${STEAMSHIP_RETVAL}` which
contains the exit code of the most recently executed command by
the user. This variable can be used by other modules to change a
section's appearance based on the most recent exit codd.