fix: add `prompt_newline` and `line_separator` prompt sections
`prompt_newline` adds a newline before the prompt is displayed,
and the display can be toggled by setting the variable
`${STEAMSHIP_PROMPT_NEWLINE_SHOW}` to `false`.

`line_separator` adds a newline into the prompt so that it spans
more than one line, and the display can be toggled by setting the
variable `${STEAMSHIP_LINE_SEPARATOR_SHOW}` to `false`.

Both variables default to `true`.