fix: change default theme and add a `bracket` theme
Change the default `steamship` theme back to using a right carat
for the prompt delimiter and a `$` or `#` for the prompt character.
This makes the theme look correct regardless of which sections are
displayed.

Add a `bracket` theme to have a connected prompt delimiter and
character, and force hiding any sections that appear between
`line_separator` and `character` in the prompt order so that the
prompt delimiter and character and stay aligned.