refactor: greatly improve the efficiency of `git_status` section
Use a `while` loop to read the output of `git status --porcelain` and
use shell globs to match various Git states within a single shell fork
instead of forking a separate `grep` to check for each Git state bit.

This reduces the number of processes that are spawned when displaying
the prompt within a Git work tree and feels much snappier.