refactor: avoid needing to specify the full module load order
Add multiple inclusion guards on each module and keep track of the order
in which modules are loaded in `${STEAMSHIP_MODULES_SOURCED}`, to be
used to determine the order in which module `init` functions are
invoked.

This avoids needing to specify the full order in which the modules need
to be loaded.