feat: add a `utils` library based on spaceship `utils.zsh`
Add a `utils.sh` library file that adds some helper functions to be used
by modules. The functions are based on similarly-named functions in the
spaceship `utils.zsh` library file.

The `upsearch` function currently differs from the spaceship one in that
it is a pure shell implementation instead of using `find`.